+++ src/texture/mod.rs
@@ -63,7 +63,7 @@ pub use image_format::{UncompressedFloatFormat, UncompressedIntFormat, Uncompres
 pub use image_format::{CompressedFormat, DepthFormat, DepthStencilFormat, StencilFormat};
 pub use image_format::{CompressedSrgbFormat, SrgbFormat};
 pub use self::any::{TextureAny, TextureAnyMipmap, Dimensions};
-pub use self::bindless::{ResidentTexture, TextureHandle};
+pub use self::bindless::{ResidentTexture, TextureHandle, BindlessTexturesNotSupportedError};
 pub use self::get_format::{InternalFormat, InternalFormatType};
 pub use self::pixel::PixelValue;