+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Added `buffer::is_buffer_read_supported`.
  - Replaced `SyncFence::new_if_supported` and `new` with `new` that returns a `Result<SyncFence, SyncNotSupportedError>`.
  - Replaced `UniformBuffer::new_if_supported` and `UniformBuffer::new` with `new` that returns a `Result<UniformBuffer, BufferCreationError>`.
+ - Replaced `ResidentTexture::new_if_supported` and `ResidentTexture::new` with `new` that returns a `Result<ResidentTexture, BindlessTexturesNotSupportedError>`.
 
 ## Version 0.7.1 (2015-07-14)